re-write deleteSection handler function in section.js controller

+++ server/controllers/Section.js
@@ -1,5 +1,6 @@
 const Section = require("../models/Section");
 const Course = require("../models/Course");
+const SubSection = require("../models/SubSection")
 
 exports.createSection = async (req, res) => {
     try{
@@ -99,24 +100,55 @@ exports.deleteSection = async (req, res) => {
     try{
 
         // Get Id- assuming that we are sending id in Params
-        const {sectionId} = req.params;
+        const {sectionId, courseId} = req.body;
 
-        // use FindByIdAndDelete
+        // Update the course and remove sectionId from course content
+        await Course.findByIdAndUpdate(courseId, {
+            $pull: {
+                courseContent: sectionId,
+            }
+        })
+
+        // Find the section
+        const section = await Section.findById(sectionId);
+        console.log(sectionId, courseId);
+
+        // Error handleing if section not found
+        if(!section) {
+            return res.status(404).json({
+                success: false,
+                message: "Section not found",
+            })
+        }
+
+        // Delete Sub-section
+        await SubSection.deleteMany({
+            _id: {
+                $in: section.subSection
+            }
+        })
         await Section.findByIdAndDelete(sectionId);
-        //TODO[Testing]: do we need to delete the entry from course schema??
+
+        // Find the updated course and return
+        const course = await Course.findById(courseId).populate({
+            path: "courseContent",
+            populate: {
+                path: "subSection"
+            }
+        }).exec();
 
         // return response
         return res.status(200).json({
             success: true,
             message: "Section Deleted Successfully",
-            // updatedCourseDetails,
+            data: course,
         })
 
     }catch(error){
+		console.error("Error deleting section:", error);
         return res.status(500).json({
             success: false,
             message: "Unable to delete section, Please Try Again",
-            error:error.message,
         });
     }
 }
